consolidated solution to use `Directory.Build.Props` options (#11)
