+++ src/Akka.Hosting/Akka.Hosting.csproj
@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <TargetFramework>$(LibraryFramework)</TargetFramework>
-	<PackageReadmeFile>README.md</PackageReadmeFile>
-    <Description>Your description here.</Description>
+		<PackageReadmeFile>README.md</PackageReadmeFile>
+    <Description>Akka.NET Microsoft.Extensions.Hosting support.</Description>
   </PropertyGroup>
 	<ItemGroup>
 		<None Include="..\..\README.md" Pack="true" PackagePath="\" />
