+++ src/Akka.Persistence.SqlServer.Hosting/Akka.Persistence.SqlServer.Hosting.csproj
@@ -1,17 +1,21 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
+   
+  <PropertyGroup>
+    <TargetFramework>$(LibraryFramework)</TargetFramework>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <Description>Akka.Persistence.SqlServer Microsoft.Extensions.Hosting support.</Description>
+  </PropertyGroup>
+  <ItemGroup>
+    <None Include="..\..\README.md" Pack="true" PackagePath="\" />
+  </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\Akka.Hosting\Akka.Hosting.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-      <PackageReference Include="Akka.Persistence.Query.Sql" Version="1.4.34" />
+      <PackageReference Include="Akka.Persistence.Query.Sql" Version="$(AkkaVersion)" />
       <PackageReference Include="Akka.Persistence.SqlServer" Version="1.4.32" />
     </ItemGroup>
