+++ src/Akka.Remote.Hosting/Akka.Remote.Hosting.csproj
@@ -1,11 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <OutputType>Library</OutputType>
-    </PropertyGroup>
+<Project Sdk="Microsoft.NET.Sdk">
+  
+  <PropertyGroup>
+    <TargetFramework>$(LibraryFramework)</TargetFramework>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <Description>Akka.Remote Microsoft.Extensions.Hosting support.</Description>
+  </PropertyGroup>
+  <ItemGroup>
+    <None Include="..\..\README.md" Pack="true" PackagePath="\" />
+  </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\Akka.Hosting\Akka.Hosting.csproj" />
